CEDS-467 Extract mocked connectors to new trait
