+++ app/connectors/CustomsDeclarationsConnector.scala
@@ -35,7 +35,7 @@ class CustomsDeclarationsConnector @Inject()(appConfig: AppConfig, httpClient: H
   def submitExportDeclaration(metaData: MetaData, badgeIdentifier: Option[String] = None)
                              (implicit hc: HeaderCarrier, ec: ExecutionContext): Future[CustomsDeclarationsResponse] =
     postMetaData(appConfig.submitImportDeclarationUri, metaData, badgeIdentifier).map{ res =>
-      Logger.debug(s"CUSTOMS_DECLARATIONS response is  --> ${res.toString} " )
+      Logger.debug(s"CUSTOMS_DECLARATIONS response is  --> ${res.toString}")
       res
     }
